23-identify-txt-records/pyproject.toml
[project]
name = "23-identify-txt-records"
version = "0.1.0"
description = ""
authors = [
    {name = "irayspace",email = "irayspacii@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
